Move reduction to voting interface
